Solve complexity issues
