+++ index.html
@@ -1,5 +1,5 @@
 
-<html lang="en"></html><head><meta charset="utf-8" /><title>ChordFiddle</title><meta name="description" content="ChordPro editor. Tranpose your chord sheet. Convert ChordPro to chord sheet." /><meta name="keywords" content="javascript, chords, parsing, convert, chord-sheet, chordpro" /><link rel="stylesheet" href="bundle.css?v=1506153783000" /></head><body><header><div class="wrapper"><h1 class="site-name"><a>ChordFiddle</a></h1><ul class="meta"><li>Version 0.2.0</li><li>Built with <a href="https://github.com/martijnversluis/ChordSheetJS" target="_blank">ChordSheetJS</a> and <a href="https://github.com/martijnversluis/ChordJS" target="_blank">ChordJS</a></li><li><a href="https://github.com/martijnversluis/chordfiddle/issues" target="_blank">Issues</a></li><li><a href="#readme">About</a></li></ul></div></header><main><section class="column column-left"><ul class="mod-toolbar"><li><button data-id="transpose-down">Transpose down</button></li><li><button data-id="transpose-up">Transpose up</button></li><li><button data-id="switch-to-sharp">Use ♯</button></li><li><button data-id="switch-to-flat">Use ♭</button></li></ul><textarea data-id="chordProEditor" class="sheet-editor">{title: Let it be}
+<html lang="en"></html><head><meta charset="utf-8" /><title>ChordFiddle</title><meta name="description" content="ChordPro editor. Tranpose your chord sheet. Convert ChordPro to chord sheet." /><meta name="keywords" content="javascript, chords, parsing, convert, chord-sheet, chordpro" /><link rel="stylesheet" href="bundle.css?v=1506627840000" /></head><body><header><div class="wrapper"><h1 class="site-name"><a>ChordFiddle</a></h1><ul class="meta"><li>Version 0.2.0</li><li>Built with <a href="https://github.com/martijnversluis/ChordSheetJS" target="_blank">ChordSheetJS</a> and <a href="https://github.com/martijnversluis/ChordJS" target="_blank">ChordJS</a></li><li><a href="https://github.com/martijnversluis/chordfiddle/issues" target="_blank">Issues</a></li><li><a href="#readme">About</a></li></ul></div></header><main><section class="column column-left"><ul class="mod-toolbar"><li><button data-id="transpose-down">Transpose down</button></li><li><button data-id="transpose-up">Transpose up</button></li><li><button data-id="switch-to-sharp">Use ♯</button></li><li><button data-id="switch-to-flat">Use ♭</button></li></ul><textarea data-id="chordProEditor" class="sheet-editor">{title: Let it be}
 {subtitle: ChordSheetJS example version}
 {Chorus}
 
@@ -32,7 +32,7 @@ future features.</p>
 <li>commit and push your changes</li>
 <li>open a pull request</li>
 </ul>
-</div></section><script src="bundle.js?v=1506153783000"></script>
+</div></section><script src="bundle.js?v=1506627840000"></script>
 <script type="text/javascript">
 //<![CDATA[
 if (window.location.hostname == 'martijnversluis.github.io') {
